% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_units.R
\name{add_units}
\alias{add_units}
\title{add_units}
\usage{
add_units(n)
}
\arguments{
\item{n}{o número a ser avaliado}
}
\value{
retorna o número com a unidade de grandeza e
duas casas decimais,
}
\description{
Esta função simplifica o valor de um número muito grande e
adiociona uma das unidades K, M, B, T.
}
\examples{
add_units(232000000)

x <- 230000000
add_units(x)

\dontrun{

add_units('2 milhões')
}

}
