% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcCCF_Pvalue.R
\name{calcCCF_Pvalue}
\alias{calcCCF_Pvalue}
\title{calcCCF_Pvalue function}
\usage{
calcCCF_Pvalue(df = ccf.df)
}
\arguments{
\item{crossCorr()}{data.frame}
}
\value{
numeric
}
\description{
Calculate crossCorr() p-value.
}
\examples{
calcCCF_Pvalue()
}
