% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSignificanceLevel.R
\name{calcSignificanceLevel}
\alias{calcSignificanceLevel}
\title{calcSignificanceLevel function}
\usage{
calcSignificanceLevel(df = SECRET.df)
}
\arguments{
\item{SECRET}{data.frame}
}
\value{
numeric
}
\description{
Calculate signficance threshold for crossCorr().
}
\examples{
calcSignificanceLevel()
}
