% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossCorr.R
\name{crossCorr}
\alias{crossCorr}
\title{crossCorr function}
\usage{
crossCorr(df = SECRET.df, plot = to.plot)
}
\arguments{
\item{SECRET}{data.frame}
}
\value{
cross correlation object (deconvolveR)
}
\description{
Calculate cross correlation for two SECRETs.
}
\examples{
crossCorr()
}
