% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curr_shares.R
\docType{data}
\name{curr_shares}
\alias{curr_shares}
\title{\enumerate{
\item Lê o arquivo com o histórico das ações
}}
\format{
Uma tabela com 12 colunas e 187 linhas:

Colunas:   id.company,
name.company,
ticker,
main.sector,
sub.sector,
segment,
listing.segment
yahoo
ref.date
qtd.ord.shares
qtd.pref.shares
total.shares

\describe{
\item{yahoo}{Código da Empresa no Yahoo Finance}
}
}
\source{
B3 e CVM
}
\usage{
curr_shares
}
\description{
\enumerate{
\item Fitra as ações correntes em relação ao último balanço
}
}
\details{
\enumerate{
\item Filtra as empresas com o índice ibr4
}
}
\keyword{datasets}
