% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors-package.R
\docType{package}
\name{factors-package}
\alias{factors}
\alias{factors-package}
\title{factors: Clean time series datasets for asset pricing factors popular in the literature.}
\description{
Various asset pricing research factor time series conveniently packaged for consumption in R.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://bautheac.github.io/factors}
  \item Report bugs at \url{https://github.com/bautheac/factors/issues}
}

}
\author{
\strong{Maintainer}: Olivier Bauthéac \email{o.bautheac@gmail.com}

}
\keyword{internal}
