% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{Fama & French}
\alias{Fama & French}
\title{Fama/French research factors}
\format{
A \linkS4class{data.table} with variables:
\itemize{
  \item{region: geographical market considered for factor construction.}
  \item{frequency: factor return frequency (day, month, year).}
  \item{period: period over which factor returns are calculated or risk free rate level is observed.}
  \item{factor: Fama/French factor}
  \item{value: corresponding return (factors) or level (risk free rate) observation.}
}
}
\source{
\url{http://mba.tuck.dartmouth.edu/pages/faculty/ken.french/data_library.html}.
}
\usage{
Fama & French
}
\description{
Return (factors) and level (risk free rate) time series for the Fama/French
  asset pricing factors.
}
\keyword{datasets}
