% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateSECRET.R
\name{generateSECRET}
\alias{generateSECRET}
\title{generateSECRET function}
\usage{
generateSECRET(df.1, df.2)
}
\arguments{
\item{df.1}{df2 data.frame with colnames as gene idenifier and one column with t-stastic}
}
\value{
data.frame
}
\description{
Generate functional probability density functions for eigenvectors
}
\examples{
generateSECRET()
}
