% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ibr4.R
\docType{data}
\name{ibr4}
\alias{ibr4}
\title{Dados das empresas viáveis da B3}
\format{
Uma tabela com 8 colunas e 182 linhas:

Colunas:   id.company,
name.company,
ticker,
main.sector,
sub.sector,
segment,
listing.segment
yahoo

\describe{
\item{yahoo}{Código da Empresa no Yahoo Finance}
}
}
\source{
B3 e CVM
}
\usage{
ibr4
}
\description{
Os dados foram gerados considerando todas empresas do índice IBrA
e além disso foram acrecentadas empresas conhecidas com boa
liquidez.
Ao todo temos 173 empresas e 182 Tickers
}
\keyword{datasets}
