% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isCCF_Significant.R
\name{isCCF_Sigificant}
\alias{isCCF_Sigificant}
\title{isCCF_Significant function}
\usage{
isCCF_Sigificant(df = ccf.df)
}
\arguments{
\item{crossCorr()}{data.frame}
}
\value{
bool
}
\description{
isCCF_Significant function
}
\examples{
isCCF_Significant()
}
