% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadGeneList.R
\name{loadGeneList}
\alias{loadGeneList}
\title{loadGeneList function}
\usage{
loadGeneList(file.IN)
}
\arguments{
\item{file.IN}{tab delimited file with three columns -  1) gene identifiers, 2) t-statistic of functional association 1, and 3) t-statistic for gene functional association 2.}
}
\value{
data.frame
}
\description{
Load tab delimited file of functional associations of gene a gene set .
}
\examples{
loadGeneList()
}
