% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrFunc.R
\name{plotCorrFunc}
\alias{plotCorrFunc}
\title{plotCorrFunc function}
\usage{
plotCorrFunc(ccf)
}
\arguments{
\item{plotCorrFunc()}{ccf object}
}
\value{
ggplot2 object
}
\description{
plotCorrFunc function
}
\examples{
plotCorrFunc()
}
